+++ src/helpers/utils/checkKeyboardLayout.ts
